TL-2 refactor code to make it more extendable
+ Renamed the classes Trade, Logger, and TradeManger to TradeImpl, LoggerImpl, and TradeManagerImpl respectively.
+ Created the interfaces Trade, Logger, and TradeManager
+ Changed all references to classes to interfaces where applicable
